Treat signed/unsigned comparison mismatches as errors
Following the -Werror=reorder precedent on the sqlite_orm interface
target: -Werror=sign-compare for GCC/Clang, /we4018 /we4388 /we4389 for
MSVC, so the mismatches reviewers kept catching by eye now fail CI. The
three existing violations - all in tests - are fixed; the library
headers compile clean under the flag.

+++ tests/operators/arithmetic_operators.cpp
@@ -31,7 +31,7 @@ TEST_CASE("Arithmetic operators") {
         for (size_t i = 0; i < rows.size(); ++i) {
             auto& row = rows[i];
             auto& name = names[i];
-            REQUIRE(int(row) == name.length() + 1000);
+            REQUIRE(size_t(row) == name.length() + 1000);
         }
     }
     {  //  +
@@ -39,7 +39,7 @@ TEST_CASE("Arithmetic operators") {
         for (size_t i = 0; i < rows.size(); ++i) {
             auto& row = rows[i];
             auto& name = names[i];
-            REQUIRE(int(std::get<0>(row)) == name.length() + 1000);
+            REQUIRE(size_t(std::get<0>(row)) == name.length() + 1000);
         }
     }
     {  //  ||
